requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
flask-migrate
python-dotenv
PyJWT
